algorithm-service/requirements.txt
grpcio
grpcio-tools
protobuf
torch
